New defaultLogConnection() and defaultResetConnection(), IDLE_READ_ONLY
New static methods defaultLogConnection(Connection, Logger) and
defaultResetConnection(Connection), which can be used to log and reset
connections in a compatible way.

New constant IDLE_READ_ONLY that defines the read-only state of
connections when reset, set to true.  Connections are reset when first
created and before being released back to the pool.  Previously, various
parts of the code would assume reset to read-only - this centralizes
that assumption.